updated post-service and message-service

+++ post-service/src/main/java/org/example/postservice/service/impl/PostServiceImpl.java
@@ -1,10 +1,11 @@
 package org.example.postservice.service.impl;
 
 import com.example.user_service.UserRepository;
-import com.example.user_service.dto.UserDto;
 import org.example.postservice.Model.Image;
 import org.example.postservice.Model.Post;
+import org.example.postservice.UserClient;
 import org.example.postservice.dto.PostDto;
+import org.example.postservice.dto.UserDto;
 import org.example.postservice.mapper.ImageMapper;
 import org.example.postservice.mapper.PostMapper;
 import org.example.postservice.repository.PostRepository;
@@ -19,9 +20,12 @@ import java.util.List;
 public class PostServiceImpl implements PostService {
     private final PostRepository postRepository;
     private final PostMapper postMapper;
-    public PostServiceImpl(PostRepository postRepository, PostMapper postMapper, ImageMapper imageMapper) {
+    private final UserClient userClient;
+    public PostServiceImpl(PostRepository postRepository, PostMapper postMapper,
+                           ImageMapper imageMapper , UserClient userClient) {
         this.postRepository = postRepository;
         this.postMapper = postMapper;
+        this.userClient = userClient;
     }
 
     public List<PostDto> findAllPosts() {
@@ -39,10 +43,12 @@ public class PostServiceImpl implements PostService {
     }
 
 
-    public PostDto createPost(PostDto postDto, MultipartFile file1, MultipartFile file2, MultipartFile file3 , UserDto userDto) {
+    public PostDto createPost(PostDto postDto, MultipartFile file1, MultipartFile file2,
+                              MultipartFile file3 , String email) {
+        UserDto userDto = userClient.getUserByEmail(email);
         Post post = postMapper.toEntity(postDto);
         if (userDto != null) {
-            post.setUserId(userDto.getId());
+            post.setUserId(post.getId());
         }
 
         List<MultipartFile> files = new ArrayList<>();
@@ -80,13 +86,22 @@ public class PostServiceImpl implements PostService {
     }
 
 
-    public void deletePostById(Long id) {
+    public void deletePostById(Long id , String email) {
+        Post post = postRepository.findById(id).orElseThrow();
+        UserDto currentUser = userClient.getUserByEmail(email);
+        if (!post.getUserId().equals(currentUser.getId())) {
+            throw new RuntimeException("Вы не владелец поста");
+        }
         postRepository.deleteById(id);
     }
 
 
-    public PostDto updatePost(Long id, PostDto postDto, MultipartFile file1, MultipartFile file2, MultipartFile file3) {
+    public PostDto updatePost(Long id, PostDto postDto, MultipartFile file1, MultipartFile file2, MultipartFile file3, String email) {
         Post post = postRepository.findById(id).orElseThrow();
+        UserDto currentUser = userClient.getUserByEmail(email);
+        if (!post.getUserId().equals(currentUser.getId())) {
+            throw new RuntimeException("Вы не можете редактировать чужой пост");
+        }
         post.setDescription(postDto.getDescription());
         post.getImages().clear();
